ImGui Context Swapping for plugins
Allows users to draw to the imgui lua bindings provided by the script API to draw to their own version of imgui.

Done by creating a Structure "ImGuiLock" that takes an ImGuiContext pointer as a parameter.

On creation it locks reframework's imgui mutex and temporarily sets REFramework's imgui context to the plugin.

On destruction it tells reframework to restore its own context from an internal pointer set on initialize, and then unlocks the imgui mutex.

This can be used